feat(odd-platform-api): get permission for query-example editing/updating (#1568)
