+++ odd-platform-ui/src/components/DataModelling/QueryExampleDetails/QueryExampleDetailsContainer.tsx
@@ -1,34 +1,29 @@
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { Box, Grid, Typography } from '@mui/material';
-import {
-  useDeleteQueryExample,
-  useGetQueryExampleDetails,
-} from 'lib/hooks/api/dataModelling/queryExamples';
+import { useGetQueryExampleDetails } from 'lib/hooks/api/dataModelling/queryExamples';
 import { useAppDateTime } from 'lib/hooks';
-import {
-  AppLoadingPage,
-  AppMenuItem,
-  AppPopover,
-  Button,
-  ConfirmationDialog,
-} from 'components/shared/elements';
-import { EditIcon, KebabIcon, TimeGapIcon } from 'components/shared/icons';
-import { useTranslation } from 'react-i18next';
-import { useNavigate } from 'react-router-dom';
-import { queryExamplesPath, useQueryExamplesRouteParams } from 'routes';
+import { AppLoadingPage } from 'components/shared/elements';
+import { fetchResourcePermissions } from 'redux/thunks';
+import { TimeGapIcon } from 'components/shared/icons';
+import { useQueryExamplesRouteParams } from 'routes';
+import { WithPermissionsProvider } from 'components/shared/contexts';
+import { Permission, PermissionResourceType } from 'generated-sources';
+import { useAppDispatch, useAppSelector } from 'redux/lib/hooks';
+import { getResourcePermissions } from 'redux/selectors';
 import QueryExampleDetailsTabs from './QueryExampleDetailsTabs';
 import QueryExampleDetailsOverview from './QueryExampleDetailsOverview';
 import QueryExampleDetailsLinkedEntities from './QueryExampleDetailsLinkedEntities';
-import QueryExampleForm from '../QueryExampleForm/QueryExampleForm';
+import { QueryExampleDetailsContainerActions } from './QueryExampleDetailsContainerActions';
 
 const QueryExampleDetailsContainer: React.FC = () => {
+  const dispatch = useAppDispatch();
   const { queryExampleId: exampleId } = useQueryExamplesRouteParams();
-  const { t } = useTranslation();
-  const navigate = useNavigate();
+  const resourcePermissions = useAppSelector(
+    getResourcePermissions(PermissionResourceType.QUERY_EXAMPLE, exampleId)
+  );
   const { data: queryExampleDetails, isLoading } = useGetQueryExampleDetails({
     exampleId,
   });
-  const { mutateAsync: deleteQueryExample } = useDeleteQueryExample();
 
   const [selectedTab, setSelectedTab] = useState(0);
   const handleTabChange = useCallback(() => {
@@ -46,13 +41,14 @@ const QueryExampleDetailsContainer: React.FC = () => {
     [queryExampleDetails?.updatedAt, formatDistanceToNowStrict]
   );
 
-  const handleDelete = useCallback(
-    async (id: number) => {
-      await deleteQueryExample({ exampleId: id });
-      navigate(queryExamplesPath());
-    },
-    [queryExampleDetails?.id]
-  );
+  useEffect(() => {
+    dispatch(
+      fetchResourcePermissions({
+        resourceId: exampleId,
+        permissionResourceType: PermissionResourceType.QUERY_EXAMPLE,
+      })
+    );
+  }, [exampleId]);
 
   return queryExampleDetails && !isLoading ? (
     <Grid container gap={2} flexDirection='column'>
@@ -63,43 +59,18 @@ const QueryExampleDetailsContainer: React.FC = () => {
           <Typography variant='body1' sx={{ ml: 1 }}>
             {updatedAt}
           </Typography>
-          <QueryExampleForm
-            btnCreateEl={
-              <Button
-                text={t('Edit')}
-                buttonType='secondary-m'
-                startIcon={<EditIcon />}
-                sx={{ ml: 1 }}
-              />
-            }
-            queryExampleDetails={queryExampleDetails}
-          />
-          <AppPopover
-            renderOpenBtn={({ onClick, ariaDescribedBy }) => (
-              <Button
-                aria-describedby={ariaDescribedBy}
-                buttonType='secondary-m'
-                icon={<KebabIcon />}
-                onClick={onClick}
-                sx={{ ml: 1 }}
+          <WithPermissionsProvider
+            allowedPermissions={[
+              Permission.QUERY_EXAMPLE_UPDATE,
+              Permission.QUERY_EXAMPLE_DELETE,
+            ]}
+            resourcePermissions={resourcePermissions}
+            render={() => (
+              <QueryExampleDetailsContainerActions
+                queryExampleDetails={queryExampleDetails}
               />
             )}
-            anchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
-            transformOrigin={{ vertical: -5, horizontal: 67 }}
-          >
-            <ConfirmationDialog
-              actionTitle={t('Are you sure you want to delete this query example?')}
-              actionName={t('Delete query example')}
-              actionText={
-                <>
-                  Query Example #{queryExampleDetails.id}{' '}
-                  {t('will be deleted permanently')}
-                </>
-              }
-              onConfirm={() => handleDelete(queryExampleDetails.id)}
-              actionBtn={<AppMenuItem>{t('Delete')}</AppMenuItem>}
-            />
-          </AppPopover>
+          />
         </Box>
       </Grid>
       <Grid item alignItems='center'>
